rename: 파일에 주어진 기능에 더 알맞게 Validation을 붙여 수정
